Normalize MR and MX default firewall rule protocol value (#249)
* Normalize MX default rule protocol case

The default rule in the MX returns with the protocol value (Any) capitalized while all user defined rules are returned with the value of protocol in lowercase

* Normalize MR default rule protocol case 

The default rule in the MR returns with the protocol value (Any) capitalized while all user defined rules are returned with the value of protocol in lowercase

* Fix typo in function name

* Fix typo in function name

* Add integration test for MX L3 Protocol case

Test that the default firewall rule protocol value is lowercase 'any'

* Add integration test for MR L3 Protocol case 

Test that the default firewall rules protocol value is lowercase 'any'

* Lowercase protocol value

Add change log fragment for returning the MR and MX L3 firewall protocol values in lowercase

* Fix integration test typo

Copy and Paste error.

* Fix test assertion

MX L3 Firewall assertion did not have the correct path

* Fix test assertion

MR L3 Firewall assertion did not have the correct path

* Normalize MR firewall rule case

The default rules in the MR return with the values for Policy, Protocol, dest_port(destPort), dest_cidr(destCidr) capitalized. This normalizes all of the values to be lowercase.

* Normalize MX firewall rule case

The default rule in the MX returns with the value for policy, protocol, src_port(srcPort), src_cidr(srcCidr), dest_port(destPort), dest_cidr(destCidr) capitalized. This normalizes all of the values to be lowercase.

* Align function name

Function normalize_case() already exists in the MX_L3_Firewall.